Moved area metric related to validation module
